Pull initials initialization into Behavior
