+++ visitz/Visitz/Behaviors/AvatarSessionBehavior.cs
@@ -1,0 +1,38 @@
+﻿using CommunityToolkit.Maui.Views;
+using Visitz.Authentication.Keycloak;
+using Visitz.Authentication.Keycloak.Events;
+
+namespace Visitz.Behaviors;
+
+internal class AvatarSessionBehavior : Behavior<AvatarView>
+{
+    private AvatarView _avatarView;
+
+    protected override async void OnAttachedTo(AvatarView bindable)
+    {
+        base.OnAttachedTo(bindable);
+        _avatarView = bindable;
+
+        await SetInitials();
+        VisitzSession.SessionChanged += VisitzSession_SessionChanged;
+    }
+
+    protected override void OnDetachingFrom(AvatarView bindable)
+    {
+        VisitzSession.SessionChanged -= VisitzSession_SessionChanged;
+
+        base.OnDetachingFrom(bindable);
+    }
+
+    private async void VisitzSession_SessionChanged(object sender, SessionChangedEventArgs e)
+    {
+        await SetInitials();
+    }
+
+    private async Task SetInitials()
+    {
+        var info = await VisitzSessionInfo.GetAsync();
+
+        _avatarView.Text = info.UserInitials ?? "--";
+    }
+}
