+++ visitz/Visitz/Views/Navigation/VerticalNavRailView.xaml
@@ -2,6 +2,7 @@
 <views:ViewModelContentView 
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:behaviors="clr-namespace:Visitz.Behaviors"
     xmlns:nav="clr-namespace:Visitz.Views.Navigation"
     xmlns:icons="clr-namespace:Visitz.FontIcons"
     xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
@@ -17,10 +18,13 @@
         <toolkit:AvatarView
             Grid.Row="0"
             Margin="0,10"
-            Text="{Binding Initials, TargetNullValue='--'}"
             BackgroundColor="{StaticResource Blue300Accent}"
             FontFamily="BCSansBold">
-            
+
+            <toolkit:AvatarView.Behaviors>
+                <behaviors:AvatarSessionBehavior />
+            </toolkit:AvatarView.Behaviors>
+
             <toolkit:AvatarView.GestureRecognizers>
                 <TapGestureRecognizer Command="{Binding OpenSessionPageCommand}" />
             </toolkit:AvatarView.GestureRecognizers>
